Add `VisitDef`.
